Add branch longer

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int [] sale = new int[] {250, 600, 150, 40};
+        long [] sale = new long[] {250, 600, 150, 40};
         SalesManager salesManager = new SalesManager(sale);
-        int max = salesManager.max();
+        long max = salesManager.max();
 
         System.out.println(max);
